+++ src/Addons/34thPRC_Overrides/config.cpp
@@ -9,6 +9,7 @@ class CfgPatches
 			"A3_Characters_F", // Vanilla
 			"DMNS_Vehicles_Bulldog", // UNSC Foundries
 			"DMNS_Vehicles_M511_Springbok", // UNSC Foundries
+			"DMNS_Weapons", // UNSC Foundries
 			"19_UNSC_Weapons", // UNSC Foundries
 			"OPTRE_UNSC_Units_Army", // OPTRE
 			"OPTRE_FC_Units_Marines", // OPTRE FC